Cleaned up UI
Arranged the UI such that when you click "Join or Create Room" it takes you to a separate screen to enter the room name and click join, with a back button to take you back to prev screen.

It doesn't really change the screen; it's just deactivating some UI Panels and activating others.

I also added some preliminary code to get the PlayerName input field text and check for null. But this may change when we start using PlayFab so I commented it out.