Fixed gun position synchronization
Forgot to mention in previous commit: I fixed the seeing through walls problem by increasing the radius of the player's collider to extend past the clipping plane of his camera.
The position of guns not owned by players were not explicitly synced. I added a Photon Transform View to all the gun prefabs and added it to the list of observed components in the gun's Photon View.
I had to remove the PTV from the PV observed components list whenever a player is holding a gun, because the gun positions are synced along with the player's PTV in that case