Made weapons vanish when wall comes down
Hacked this solution a bit. There was a fatal error created in an edge case using PhotonNetwork.Destroy() on the guns. So, for now, I'm just moving the guns out of the reach of the players. To them, it looks the same as if they were being completely removed from the game.