Gun position synced
Finally! No bugs/glitches/whoopsies! Not how I originally wanted to implement this where a gun with a photon view would be added to a player's hierarchy and and either the gun's PV or the player's PV would synchronize position of gun. Nothing I tried worked... And I tried everything!

Current implementation is a little more confusing but requires no additional network traffic to "sync" gun positions.
This is how I do it:
When a gun (with a PV) is picked up, it is placed in the player's GO hierarchy as a child of "Inactive Weapons" GO and is disabled (no longer visible).

When a gun is selected to become the "active gun", it is moved in the player's GO hierarchy to be a child of "Active Weapon" GO. Here's the funky part -> *It is kept disabled (invisible)* and a second gun prefab is instantiated *locally* (i.e., not on the network. i.e., not explicitly synced with photon). You'll see me refer to this as the "show" gun in the code. The show gun prefab that we instantiate is chosen based on the type of "active" gun (the networked gun that we picked up and eventually put in "Active Weapon" GO). The show gun is the one used for displaying to the user and shooting and all that jazz.
When the active gun is dropped:
1) the show gun is destroyed
2) the active gun is re-enabled and tossed into the arena for other players to pick up again (the position of the active gun is synced with its own photon view and photon transform view components)

I also had to do some work figuring out how to get clients who joined the room after players had picked up guns to set up their players with the same guns. I won't go into those details because I doubt anyone will need/care to know