Team Balancing Bug Fix
When two players would join at the same time, the second player would think there was no one else in the room and join Team A (where the first player was) and set the wrong size for Team A in Room.CustomProperties. To fix this I only allow the master client to instantiate a local player without checking what team they should be on. I created a couple of Events and made anyone who was not the master client raise an event on the master client asking the mc to decide what team they should be on. The mc decides and raises an event on the client who asked for the info. The client then instantiates their player. There is one known bug that I will fix later (it shouldn't affect anyone noticeably in the meantime). There is another problem with this solution I have to figure out. What happens if the second client asks the master client for information but the master client quits before responding? I'll figure that out later as well, and again, I don't assume this bug will impact you guys either.