Extend Wall Timer when Gun is picked up
Now the wall timer will be extended if a gun is picked up during stage 1 at a weapon spawn point. It was important to make sure the gun was at a spawn point to prevent players dropping the gun and picking it back up repeatedly to extend the timer forever
I hard-coded a 5 second extension but this can be easily changed in the future if we want. We might want to take into account how many players are playing and/or how many guns are available to be picked up.