Lots of Smart AI changes
Still buggy
The goal was to make the AI player look for guns in stage 1 by traveling to locations based on the list of guns spawned on his teams side. While traveling, if an unclaimed gun came into view and that gun was not the gun we were targeting, the AI would change course and try to pick up that gun first. This is so much more complicated than I had expected.

I also had to modify the guns because this was the first time I noticed how screwed up they were. For example, there were multiple photon views on at least one of the guns. Not good! Also had to make the colliders for the guns a little bigger to ensure the AI player would hit them.