Toggle AI improvements
Now the top panel changes color to make it clear that AI control is enabled.
Also, the player must be grounded (on a nav mesh) to enable AI so I implemented that check. Nothing happens if you jump in the air and hit T to toggle AI. Before it would break things and you'd freeze in the air if you hit T while jumping.