Put weapon spawn points in both red and blue bases. Also attempting some fixes on some bugs i noticed. Lots to do
Currently having some issues with the playergrenade icons. if i try and check for grenadeicons[0] when we endround() it generates a nullref for some reason.

Also having a slightly bigger bug with the remote client not dropping space guns at the end of a round.

Also I finally figured out why the round doesnt roll over sometimes-- it's just due to a nullref.

Another prevalent issue to fix before the demo is nullrefs from aigunwatcher. This will help our program flow better